+++ example/lib/conditional.dart
@@ -29,33 +29,37 @@ class App extends StatelessWidget {
   final _loggedInRoutes = [
     GoRoute(
       pattern: '/',
-      builder: (context, state) => MaterialPage<FamiliesPage>(
-        key: const ValueKey('FamiliesPage'),
-        child: FamiliesPage(families: Families.data),
+      builder: (context, state) => MaterialPage<HomePage>(
+        key: const ValueKey('HomePage'),
+        child: HomePage(families: Families.data),
       ),
-    ),
-    GoRoute(
-      pattern: '/family/:fid',
-      builder: (context, state) {
-        final family = Families.family(state.params['fid']!);
+      routes: [
+        GoRoute(
+          pattern: 'family/:fid',
+          builder: (context, state) {
+            final family = Families.family(state.params['fid']!);
 
-        return MaterialPage<FamilyPage>(
-          key: ValueKey(family),
-          child: FamilyPage(family: family),
-        );
-      },
-    ),
-    GoRoute(
-      pattern: '/family/:fid/person/:pid',
-      builder: (context, state) {
-        final family = Families.family(state.params['fid']!);
-        final person = family.person(state.params['pid']!);
+            return MaterialPage<FamilyPage>(
+              key: ValueKey(family),
+              child: FamilyPage(family: family),
+            );
+          },
+          routes: [
+            GoRoute(
+              pattern: 'person/:pid',
+              builder: (context, state) {
+                final family = Families.family(state.params['fid']!);
+                final person = family.person(state.params['pid']!);
 
-        return MaterialPage<PersonPage>(
-          key: ValueKey(person),
-          child: PersonPage(family: family, person: person),
-        );
-      },
+                return MaterialPage<PersonPage>(
+                  key: ValueKey(person),
+                  child: PersonPage(family: family, person: person),
+                );
+              },
+            ),
+          ],
+        ),
+      ],
     ),
   ];
 
@@ -75,8 +79,8 @@ class App extends StatelessWidget {
       context.watch<LoginInfo>().loggedIn ? _loggedInRoutes : _loggedOutRoutes;
 
   Page<dynamic> _errorBuilder(BuildContext context, GoRouterState state) =>
-      MaterialPage<Four04Page>(
-        key: const ValueKey('Four04Page'),
-        child: Four04Page(message: state.error.toString()),
+      MaterialPage<ErrorPage>(
+        key: const ValueKey('ErrorPage'),
+        child: ErrorPage(message: state.error.toString()),
       );
 }