+++ example/lib/url_strategy.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
 
-import 'shared/data.dart';
 import 'shared/pages.dart';
 
 void main() {
@@ -14,7 +13,7 @@ void main() {
   runApp(App());
 }
 
-/// sample app using the path URL strategy, i.e. no # in the URL path
+/// sample class using simple declarative routes
 class App extends StatelessWidget {
   App({Key? key}) : super(key: key);
 
@@ -22,7 +21,7 @@ class App extends StatelessWidget {
   Widget build(BuildContext context) => MaterialApp.router(
         routeInformationParser: _router.routeInformationParser,
         routerDelegate: _router.routerDelegate,
-        title: 'URL Strategy GoRouter Example',
+        title: 'Declarative Routes GoRouter Example',
       );
 
   late final _router = GoRouter(
@@ -35,39 +34,23 @@ class App extends StatelessWidget {
   List<GoRoute> _routesBuilder(BuildContext context, String location) => [
         GoRoute(
           pattern: '/',
-          builder: (context, state) => MaterialPage<FamiliesPage>(
-            key: const ValueKey('FamiliesPage'),
-            child: FamiliesPage(families: Families.data),
+          builder: (context, state) => const MaterialPage<Page1Page>(
+            key: ValueKey('HomePage'),
+            child: Page1Page(),
           ),
         ),
         GoRoute(
-          pattern: '/family/:fid',
-          builder: (context, state) {
-            final family = Families.family(state.params['fid']!);
-
-            return MaterialPage<FamilyPage>(
-              key: ValueKey(family),
-              child: FamilyPage(family: family),
-            );
-          },
-        ),
-        GoRoute(
-          pattern: '/family/:fid/person/:pid',
-          builder: (context, state) {
-            final family = Families.family(state.params['fid']!);
-            final person = family.person(state.params['pid']!);
-
-            return MaterialPage<PersonPage>(
-              key: ValueKey(person),
-              child: PersonPage(family: family, person: person),
-            );
-          },
+          pattern: '/page2',
+          builder: (context, state) => const MaterialPage<Page2Page>(
+            key: ValueKey('Page2Page'),
+            child: Page2Page(),
+          ),
         ),
       ];
 
   Page<dynamic> _errorBuilder(BuildContext context, GoRouterState state) =>
-      MaterialPage<Four04Page>(
-        key: const ValueKey('Four04Page'),
-        child: Four04Page(message: state.error.toString()),
+      MaterialPage<ErrorPage>(
+        key: const ValueKey('ErrorPage'),
+        child: ErrorPage(message: state.error.toString()),
       );
 }